Reorganize servlets template
