--- web/src/main/java/com/infoshareacademy/gitLoopersi/servlet/FindHolidayByDateServlet.java
@@ -1,23 +1,0 @@
-package com.infoshareacademy.gitLoopersi.servlet;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@WebServlet("/find-holiday-by-date")
-public class FindHolidayByDateServlet extends HttpServlet {
-
-  private Logger logger = LoggerFactory.getLogger(getClass().getName());
-
-  @Override
-  protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-      throws ServletException, IOException {
-
-  }
-
-}
