+++ web/src/main/java/com/infoshareacademy/gitLoopersi/servlet/HomeServlet.java
@@ -9,8 +9,8 @@ import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@WebServlet("/json-load")
-public class LoadFromJSONServlet extends HttpServlet {
+@WebServlet("/home")
+public class HomeServlet extends HttpServlet {
 
   private Logger logger = LoggerFactory.getLogger(getClass().getName());
