+++ web/src/main/java/com/infoshareacademy/gitLoopersi/servlet/admin/AddDayOffServlet.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.gitLoopersi.servlet;
+package com.infoshareacademy.gitLoopersi.servlet.admin;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -9,8 +9,8 @@ import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@WebServlet("/additional-vacation")
-public class AdditionalVacationServlet extends HttpServlet {
+@WebServlet("/admin/vacation/add-day-off")
+public class AddDayOffServlet extends HttpServlet {
 
   private Logger logger = LoggerFactory.getLogger(getClass().getName());
